Attempted to find the correct time by a modified LCM approach.
The approach in https://en.wikipedia.org/wiki/Least_common_multiple#Using_a_simple_algorithm
gives a way to find the LCM of a _set_ of numbers, which is useful here.

It's possible to subtract the time offsets from each of the buses at the
start, so instead of looking for T such that
`T % b1 == 0 && (T+1) % b2 == 0 && ...`

Instead we're looking for `T == bootleg_lcm([b1, b2-1, b3-3, ...], [b1,
b2, b3, ...])`, incrementing beginning numbers by the bus number.

This worked great for the example sets! But it's... not super fast. I
let it run on my i7-1165G7 for several minutes, and while it _did_ spend
about 10% of my battery, it did not find a solution.

Back to the drawing board. I was going to try a Euclidean Algorithm
approach in hopes that that'd perform a bit better (see code), but I was
having a hell of a time figuring out how to factor the offsets in.

Then I cheated and got a hint from Reddit. So I'm going to do that.